Extract support switch and location check into support-utilities
