+++ static/src/javascripts/projects/common/modules/commercial/membership-engagement-banner.js
@@ -15,7 +15,7 @@ import {
     submitComponentEvent,
     addTrackingCodesToUrl,
 } from 'common/modules/commercial/acquisitions-ophan';
-import { selectBaseUrl } from 'common/modules/commercial/contributions-utilities';
+import { selectBaseUrl } from 'common/modules/commercial/support-utilities';
 
 // change messageCode to force redisplay of the message to users who already closed it.
 const messageCode = 'engagement-banner-2017-09-21';
